Refactoring & Last touches

+++ src/main/java/com/dream_tournament/dto/GroupLeaderboardEntry.java
@@ -1,7 +1,7 @@
 package com.dream_tournament.dto;
 
 public record GroupLeaderboardEntry(
-        Long userId,
+        Integer userId,
         String username,
         String country,
         Integer score
